requirements.txt
asgiref==3.11.1
Django==6.0.5
djangorestframework==3.17.1
gunicorn==26.0.0
packaging==26.2
sqlparse==0.5.5
whitenoise==6.12.0
